lostfound_backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.3.0

# File Upload
python-multipart==0.0.6

# Environment Variables
python-dotenv==1.0.0
